[Bugfix:InstructorUI] Fix Histogram Statistics & Rendering (#11729)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

The manual grading histogram standard deviation for gradable statistics
does not work as it always displays `NaN` given the `manual_average`
variable on the backend should be of type `SimpleStat`, such as
`overall_average`, but the current approach implements it as a sum of
all component averages. Additionally, when selecting a new bucket size
there is an inconsistency where the `Range` statistic appears and the
`Standard Deviation` statistic disappears.


![image](https://github.com/user-attachments/assets/db14d960-d5c4-4c56-a028-18bce741b772)

![image](https://github.com/user-attachments/assets/6e096c1b-4e68-418e-8146-2055616035d7)


### What is the new behavior?

The standard deviation calculation has been fixed to display the proper
statistics based on the overall manual grading data. Additionally, for
assignments with missing histogram/scatterplot statistics, instead of
not rendering the respective tab, a missing data message is displayed
for simplicity.


![image](https://github.com/user-attachments/assets/c76c5bcf-64cd-4882-b5b8-704c3c2670f9)

![image](https://github.com/user-attachments/assets/24f0f00e-2023-41c0-87f2-a99071d78e8a)



### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

Fixes #10944 and #10864.