[Bugfix:InstructorUI] DockerUI type mismatch (#11739)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
This change is necessary because it fixes a type mismatch that causes a
crash in the DockerUI page. WorkerMachineSystemInformation.php contains
nullable strings for shipper and daemon service statuses but are not
nullable in the constructor. When sysinfo logs of shipper or daemon
services are "Service Not Found" or missing, DockerUI.php sets these
values to null, causing the mismatch.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
DockerUI page should not show frog robot when worker, shipper, or daemon
services are null in DockerUI.php. The status of the services are not
nullable (fallback to "No Status Collected"). I have also made it so
that if the values are set to null or "Service Not Found", rather than
not displaying the entry at all, it displays the status of the service.
If they are missing in the sysinfo.log, it just shows "No Status
Collected".

Before

![image](https://github.com/user-attachments/assets/78889b74-7680-45b9-b0dd-54d1be8b4009)

After

![image](https://github.com/user-attachments/assets/ec4f9cec-f01c-40cb-a248-6463c6e84ea3)

### What steps should a reviewer take to reproduce or test the bug or
new feature?
To reproduce, you can manually modify the sysinfo log in
/var/local/submitty/logs/sysinfo/YYYYMMDD.txt to remove shipper or
daemon, or have their status as "Service Not Found". Then refresh the
DockerUI page and it will result in this crash.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
I have added PHP unit testing to test whether the worker, shipper, or
daemon services are correctly set to null and their respective status in
the WorkerMachineSystemInformation.php.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
I have also fixed some minor inconsistencies in style in code.