[Refactor:TAGrading] Refactor clear version conflict test (#11722)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
The clearing_version_conflicts.spec.js Cypress test directly calls
functions from the client js of TAGrading which means that any changes
to that will break this test.

### What is the new behavior?
The test has been updated to make it more independent and now uses
`cy.request` to make the changes it needs.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
This is needed before #11716 since that changes the js too much for
cypress to pass.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>