[Bugfix:InstructorUI] Correct machine on docker error (#11744)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Fixes #11725
Currently, the value for the machine incorrectly displays the container
as the machine. We want users/instructors to know which exact machine is
missing the images.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Before

![452027974-98242ab0-87d9-414d-8f7f-efac3c5feb19](https://github.com/user-attachments/assets/bb285a9b-5a88-41b4-9345-4048d4b38d1b)

After

![image](https://github.com/user-attachments/assets/1bae183f-298f-47c4-a543-57e891d2f0c9)

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Submit example submissions to a course where you don't have a docker
image for. For example, Docker Choice Of Language uses rust:latest,
clang:latest, gcc:latest.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
Cypress tests have not been created yet for autograding, it is in
progress.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>