Remove Spellcheck, Unused var, and added auto scrolling
QOL for the feature. There were many spellcheck suggestions in the config files that were unnecessary.

Additionally, the user had to scroll to the bottom after opening the text editor, so I added automatic scrolling which overall improves the experience. I chose 800px because I used the same function in both instances, and wanted to ensure that it would go to the absolute bottom when opening the editor, but stay near the top of the text area when that is opened.

Known Issue: When the user selects an autograding configuration, they have to reload the page upon successful build completion to be able to open the editor.