Fixed Future Dates being displayed
This was a one-off issue where if today is 06/16, all tags set for 06/17 would be displayed, but nothing from 06/18 and beyond.

To fix, I altered the date logic to only compare days, rather than hours (which was the problem). This resulted in 0 being returned when the category was set for the present date, so I updated all of the logic to be ">= 0"