[Bugfix:Submission] Move Queue Text to global AutoResults (#11747)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Currently the user has two different texts that show them the same text
that they are somewhere in the queue. This is because the instructor
view has the queue in addition to the student queue. If the submission
is currently getting graded, then we can call the server to check for
results, which was not happening
Impacting #11703 

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
The code only appears once, in addition to currently grading checking
for the refresh.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Submit a gradeable
Refresh until you see it is being graded
page would not refresh

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
N/A 

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
I added a delay to autoresultsbox because there is a race condition that
there might be a status if we press submit.
#5899 explains it better. The delay was added because we might refresh 3
times very quickly and not catch the new queue state. I thought 1.5
seconds was a reasonable amount of time.