[Refactor:Developer] Use random.choices in Python utils (#11720)
Please check if the PR fulfils these requirements:
- [ ] Tests for the changes have been added/updated (if possible)
- [ ] Documentation has been updated/added if relevant
- [ ] Screenshots are attached to the GitHub PR if visual/UI changes
were made

### What is the current behaviour?
The utility function currently uses a generator for sampling/choosing
from a population.

### What is the new behaviour?

This PR will optimise the utility function for more effective batch
processing by using `random.choices` with the specified length
parameter(`k`). It is preferable over looping.

### Other information?

The relevant tests already exist for the set of utilities, and hence
this PR doesn't require additional tests to be written. Although the
change/patch may seem minor due to a single call in the autograder
module (though this one is more efficient), the code becomes more
concise and readable as well as more _Pythonic_.