[Bugfix:InstructorUI] Syllabus Bucket Validation Checks (#11544)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?

Instructors may assign incorrect or arbitrary values to the assignment
syllabus bucket within the creation and/or update request, which is an
issue as when they do provide an invalid syllabus bucket, the grades
configuration page triggers a site error, causing Rainbow Grades to
break. There is a clear error in the upload gradable JSON example,
`/site/cypress/fixtures/json_ui.json`, where the syllabus bucket for
homework is capitalized, but the backend expects the value to be
lowercase.


![image](https://github.com/user-attachments/assets/5bdf671e-f79d-4ea9-98ae-0e658b307e8b)


### What is the new behavior?

The backend validates the syllabus bucket both during initial creation
and when updates are made through the updates route. If an invalid
syllabus bucket is detected, an error message is displayed to the
instructor. Additionally, modifying the input value of the syllabus
bucket on the front end, such as through DOM manipulation, triggers a
save error.


![image](https://github.com/user-attachments/assets/4cb81f90-87a6-4b44-9ab8-8ca5815b860d)


### Other information?

To test, upload a gradable through the JSON feature (located at the top
right of the construction page) using the existing example in
`site/cypress/fixtures/json_ui.json`. This should trigger an error on
the current branch while being successful on the main branch, which
addresses the broken Grades Configuration page. This fixes the invalid
syllabus bucket example introduced in #10100.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>