[Refactor:InstructorUI] Refactor SQL Queries to use vue (#11738)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
For #11643, it would be better to update the queries list instead of
fetching it everytime. This PR would faciliate that, while making the
other PR look smaller in comparison

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->

Old:

https://github.com/user-attachments/assets/0f174827-4676-4290-ac57-ad891a98f176
New:

https://github.com/user-attachments/assets/10c12139-1399-4b37-8795-f887bb93f7f1
New Video will be posted later

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Make sure all things work, click buttons on that page.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
I had to delete the js unit tests because I removed a lot from the TS
file and would not be accurate to accurately replicate it in the unit
tests. E2e issue will be made. There are units tests testing this
feature, which should accurately replace the js unit test

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->