[Testing:Developer] Fix Cypress Apache Fails (#11724)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Currently cypress tests are flaky due to apache failing to start. The
reason that was found due to testing was because nginx was using port 80
as part of the default install.

### What is the new behavior?
This installs **all** of the nginx stuff we need before we install
apache. Since nginx is served on port 8443, we should not get an apache
fail error.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

---------

Co-authored-by: Williams Chen <82904790+williamschen23@users.noreply.github.com>
Co-authored-by: williamschen23 <chenwill2005@gmail.com>