[Refactor:TAGrading] Refactor and Modularize TAGrading (#11716)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
TAGrading is mostly split between ta-grading.ts (2500 lines) and
ta-grading-rubric.ts (3500 lines) and since it has recently been
migrated from javascript there is a lot of room to improve the code
quality and readability.

### What is the new behavior?
This PR moves some behavior to move off of the window element and uses
export instead. In addition, there is a couple lines of changes that
fixes the notebook view, which is currently broken when you are
initializing the page with a two panel design. More information can be
found because this fixes #11740

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

---------

Co-authored-by: williamschen23 <chenwill2005@gmail.com>
Co-authored-by: Williams Chen <82904790+williamschen23@users.noreply.github.com>