[Feature:Notifications] Released grade notification (#10358)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently, students are not notified when a score for a visible gradable
is released. It is up to the instructor to release an announcement for
the given assignment.

### What is the new behavior?
Now, all users who wish can receive notifications and/or emails
regarding all released scores for any visible electronic gradeable
scores for the course they are registered in. Notification settings can
be updated as usual in the notification settings page to toggle these
notifications as users please.


![image](https://github.com/user-attachments/assets/cc374e54-4004-47d0-bfd5-0504e4211114)

 
Instructors can now also view the total notifications sent within the
dates section. If they wish to resend them for a future release date, a
confirmation is displayed on the screen when changing to a future date
for gradeables with released notifications.


![image](https://github.com/user-attachments/assets/4eeb5680-d079-4d58-979c-ad47254a948f)

### Other information?
This is implemented as a cron job, a scheduled task, to run every hour.
To test this feature, you must run the course migration as the
notification state of the gradable is added to the gradable table. You
can initially run `submitty_install` to run the proper migration. From
here, you can run the following commands after ensuring at least one
gradable has pending notifications.


```bash
$ sudo su submitty_daemon
$ python3 /usr/local/submitty/sbin/send_notification.py
$ cat /var/local/submitty/logs/notifications/20250319.txt
   [2025-03-21 17:12:14.007012] Successfully updated notification status for 0 submissions

   [2025-03-21 17:14:40.049533] (sample): Sent 81 site, 79 email notifications
   [2025-03-21 17:14:40.101745] Successfully updated notification status for 83 submissions

   [2025-03-21 17:24:45.058778] (sample): Sent 46 site, 45 email notifications
   [2025-03-21 17:24:45.095482] Successfully updated notification status for 46 submissions
```

To re-release notifications as an instructor for an existing gradable
with released notifications, a confirmation will pop up on the edit
dates page if the release date is being changed to a future date, where
`OK` should be clicked to reset the notification status of recipients.

```
[Submitty sample] Grade Released: Grading Homework PDF
Your grade is now available for Grades Released Homework in course 
SAMPLE.

Click here for more info: http://localhost:1511/courses/s25/sample/gradeable/grading_homework_pdf

--
NOTE: This is an automated email notification, which is unable to receive replies.
Please refer to the course syllabus for contact information for your teaching staff.
Update your email notification settings for this course here: http://localhost:1511/courses/s25/sample/notifications/settings
```


![image](https://github.com/user-attachments/assets/a9f46e35-6851-4556-b4d6-0833b32d4be3)

---------

Co-authored-by: Jeffrey <94331560+jeffrey-asm@users.noreply.github.com>
Co-authored-by: Zheyu Deng <122962727+ZheyuDeng@users.noreply.github.com>
Co-authored-by: Zheyu Deng <sony1119266216@gmail.com>
Co-authored-by: Justin Manion <jmanion32@gmail.com>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Barb Cutler <Barb Cutler>