[Feature:InstructorUI] Capability / image buildtime error (#10839)
Works on #10747 
Gives an error for gradeables at buildtime if the requested capability /
image does not exist in the context of that gradeable.


![image](https://github.com/user-attachments/assets/65b4dc2d-fd83-4023-9dba-711529197f4f)


![image](https://github.com/user-attachments/assets/ee83ae00-f320-4a76-af8d-0446ee1fb76c)

To test, you can modify an existing config for a development course to
either have a capability that does not exist or an image that does not
exist and rebuild the gradeable for that course.

---------

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>
Co-authored-by: Christopher Poon <christopherpoonc@gmail.com>
Co-authored-by: williamschen23 <chenwill2005@gmail.com>