Fixed Diff, Removed Restrictions for Filter
Filter withdrawn students was set to be instructor only, but should be all non-student graders. This was causing Cypress to fail on page load since we check if each box is checked.

This will likely need further improvement to validate permissions on reload so we never check for a box that is not visible on the page.