Enhanced Permissions for Checkboxes
The function that ensures the right checkboxes are toggled on page load assumed all graders had the same permissions, reading to reference errors, and failing Cypress tests. Now, the function considers these permissions.

Additionally, the withdrawn students toggle is no longer displayed for team work. Instead, it is set to on for everyone.