[Bugfix:TAGrading] Always show new columns in TAGrading (#11794)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
When switching between different TAGrading views, there is a regression
in #11487 that means non-default columns will be hidden by default.
Sometimes it does not even show toggles to enable the new columns.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
The list of columns is dynamic and automatically enabled for any columns
that have not yet been seen by the user.
Fix #11754 

### What steps should a reviewer take to reproduce or test the bug or
new feature?
On main:
1. Open the grading details page
2. Turn off first and last name
3. switch to anon mode

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->