[Bugfix:TAGrading] Mark component as modified (#11785)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Due to setting `Component::modified` inside:

https://github.com/Submitty/Submitty/blob/bcdd81478559d0c66552e2d494e91878b5c8a79a/site/app/models/AbstractModel.php#L131-L207
As we are removing the magic methods we have to make sure that we also
set `modified` in all of the setters.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
All of the setters that did not update `Component::modified` now do.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
There is no way to currently produce incorrect behavior, since all code
paths that lead to `saveComponent` either call it without the
possibility of the component being edited or they call an existing
`Component` setter as well as the new ones so `Component::modified` is
already set to true.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->