[Refactor:TAGrading] Remove Grading Beta (#11800)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
`grading_beta` was introduced in #9454 and improved on in #9483 . Popup
refactor was worked on in PRs
https://github.com/Submitty/Submitty/pull/7407,
https://github.com/Submitty/Submitty/pull/7812, and
https://github.com/Submitty/Submitty/pull/7847. We currently do not have
any developers working on these issues for around 3 years now, and it
has become obsolete. As the current TAGrading is being refactored in
vue, they will become even more obsolete. Since it is in a unmainted
statement, I propse that we remove it completely so that we don't have
to change `popup_refactor` in the current code

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Normal Behavior of the site should not change.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->