[Bugfix:InstructorUI] DockerUI bug fixes (#11804)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Fixes #11749
This change is necessary because it fixes up bugs in DockerUI.
Currently, there are duplicate image entries if you have multiple worker
machines, which expand the page and make it hard to view. In addition,
system info is incorrectly parsed causing the OS and docker version for
each machine to fallback to "Unknown". I also changed the ID of the
image to display its corresponding digest on DockerHub.

### What is the New Behavior?
Images should not be duplicated, system info for OS and docker version
should not be unknown, and digest is shown instead of ID.

Note that the digest reflects the index digest (overarching digest for
multi-architecture bundled images). In our case, since some images
support arm and amd, the index digest will be displayed. If there is
just one supported architecture, the index digest reflects the manifest
digest.

<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Before

![image](https://github.com/user-attachments/assets/87da6d98-e1a8-45e4-b7db-70f4db1b0de4)

![image](https://github.com/user-attachments/assets/5f9112bb-1fb6-4509-b2b1-0d4598f97ce3)

After

![image](https://github.com/user-attachments/assets/50e58ff9-61cc-43b2-91ec-a88b0154c6ee)

![image](https://github.com/user-attachments/assets/546b7e81-9ffb-4a79-847b-dc4197707b6f)

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
I updated the relevant DockerUI cypress test. 

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>