Fixed JS Lint and Cypress, Patched Dup Bug
When the user has multiple files that they can toggle between, it was possible to infinitely duplicate the CodeMirror text areas since they didn't close on the original on change. Now, when the user toggles between 2 files, it closes the original file. The only issue is there is a frame where the open text area closes and the new one opens, which is slightly abrupt for the user experience, but does not affect functionality.