[Bugfix:Notifications] Fix Invalid Email Status Pagination (#11806)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->

The email status page only displays the first page, regardless of the
currently selected page. I believe this was never caught or documented,
as most instructors or administrators would usually visit this page to
verify recent, but not old, email details. Fixes #11799.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->

The email status page now displays the correct page. The error stemmed
from the invalid controller logic, where the page was determined by the
POST request object, which was empty or undefined, as the pagination
request is a GET request. Additionally, the repository code was
refactored for clarity measures.

Old (all pages are the same)


![image](https://github.com/user-attachments/assets/5fbb8b30-b182-4e9e-83ff-21932d0bac0c)


New (expected last page)


![image](https://github.com/user-attachments/assets/187b3086-b75f-4b20-85ac-9f30758d340a)

### What steps should a reviewer take to reproduce or test the bug or
new feature?

You can test the expected pages (10 emails or 5000 recipients per page)
through the following commands, which are based on emails, not recipient
limits. The following is also not tied to a specific course, which would
require the `WHERE course='x' AND term='y'` clause to work, thus log in
as the `superuser` to match the emails.

```bash
$ sudo su postgres
$ psql
$ \c submitty
$ SELECT
  ROW_NUMBER() OVER (ORDER BY e.created DESC) AS index,
  e.subject,
  e.created,
  COUNT(*) AS cnt
FROM emails e
GROUP BY e.subject, e.created
ORDER BY e.created DESC;
```


### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

There are no existing email status page tests, but I'll make an issue to
ensure they're implemented as E2E tests to account for pagination,
errors, and detail verification.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

N/A