[Feature:System] Added Sorting for Pending Gradeables (#11753)
### Why is this Change Important & Necessary?
fixes #11748
The table on the pending gradeables page has no sort options. This is
dangerous because admins should look at this page to ensure they don't
install when a course is nearing a deadline.

### What is the New Behavior?
The table can now be sorted by course (with gradeable as the tiebreaker)
and by due date.
![sorted pending gradeables in ascending
order](https://github.com/user-attachments/assets/22962043-ace0-4719-bf40-93538d71c87f)

### Other information
This is not a breaking change. This PR breaks the sortable_header macro
into its own file and adjusts docker.twig and gradeables.twig to both
use it.

---------

Co-authored-by: Alex Lavallee <alex@lavallee.one>