[Feature:InstructorUI] Add CodeMirror to Gradeable Editor (#11814)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
When editing a gradeable config, using `TAB` to indent is not supported.
This hinders the experience of instructors editing large config files,
making it difficult to maintain consistent indentation.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
The text area used to edit the gradeable config is now CodeMirror. This
allows the use of `TAB` to make indentations, while having a working
undo/redo system.

I chose CodeMirror over a standard JavaScript solution because while I
can insert a tab with something like a `keydown` event listener, the
undo/redo functionality of the DOM gets messed up. CodeMirror offers
built-in undo/redo tracking and keyboard handling, making it a more
robust and accessible solution compared to patching native `<textarea>`
behavior with `keydown` events.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Go to a gradeable with an uploaded config file -> open editor -> observe
that `TAB` is now supported, and all other functionality is the same.

### Other information
<!-- Is this a breaking change?  
This is not a breaking change.
A few `!important` declarations were added to override CodeMirror's
internal styles and align the config editor with existing Submitty
design patterns. These follow the precedent set in
`gradeable-notebook.css`.